Update docker.md
This doc suggested using this to install Laravel:
```
curl -s "https://laravel.build/example-app" | bash
```
But that currently installs Laravel 10, which is incompatible with Statamic. This will likely be a recurring problem since Statamic will lag behind major releases of Laravel. While that command is convenient it might be best not to mention it for this reason.